Make regression UQ robust to missing smiles
fit_regression_uq now treats prox_df as optional: V0 (calibration, no
structure needed) is always fit; the fingerprint-proximity V1/V2 are fit
only when a smiles-backed prox_df is provided, with a v0 fallback if a
fingerprint version is requested without one. The xgb/pytorch/chemprop
templates pass prox_df=None (with a log line) when no smiles column is
present instead of hard-failing, so non-molecular regressors can train.
